+++ src/insertion.rs
@@ -39,8 +39,21 @@ impl Insertion {
         self.text.range.end
     }
 
-    #[inline(always)]
-    pub(crate) fn inserted_by(&self) -> ReplicaId {
+    /// Returns the [`ReplicaId`] of the [`Replica`] that performed the
+    /// insertion.
+    ///
+    /// # Examples
+    ///
+    /// ```
+    /// # use cola::Replica;
+    /// let mut replica1 = Replica::new(1, 3);
+    ///
+    /// let insertion = replica1.inserted(3, 7);
+    ///
+    /// assert_eq!(deletion.deleted_by(), replica1.id());
+    /// ```
+    #[inline]
+    pub fn inserted_by(&self) -> ReplicaId {
         self.text.inserted_by()
     }